stabilize tests
